pyproject.toml
[project]
name = "msft-hackaton-gitllmexplorer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "gitpython>=3.1.44",
    "openai-agents>=0.0.11",
]
